requirements.txt
dominate
visdom
matplotlib
numpy
python-dateutil
pytz
PyYAML
torch==1.8.0
torchvision